backend/pyproject.toml
[project]
name = "roam-backend"
version = "0.1.0"
description = "Roam semantic sync backend services"
requires-python = ">=3.11"
dependencies = [
    "aiolimiter>=1.2.1",
    "chonkie[semantic,voyageai]>=1.3.1",
    "fastapi>=0.118.0",
    "funcy>=2.0",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.1",
    "sqlalchemy>=2.0.43",
    "structlog>=25.4.0",
    "tenacity>=9.1.2",
    "tqdm>=4.66",
    "typer>=0.19.2",
    "uvicorn[standard]>=0.37.0",
    "voyageai>=0.3.5",
    "weaviate-client>=4.17.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
